+++ NEWS.md
@@ -1,3 +1,16 @@
+# rbi 0.8.0
+
+* `supplement` option to `save_libbi` to store supplementary information
+* Temporary directories are now removed when quitting R or collecting garbage
+* Improved chaining of `libbi` calls
+* New `attach_file` function, a more general version of the former `add_output`
+* New `file` option for `bi_read`
+* Fixed problems with `predict`
+* Fixed add_output when passed a list
+* Fixed a bug in writing observation files; extra dimensions are now written in the correct order
+* Improved writing of observation files with `ns` or other dimensions
+* Improved error handling: `libbi` calls now finish and yield a warning if the call to the libbi executable fails; this makes it possible to inspect the object to diagnoise the problem
+
 # rbi 0.7.0
 
 * The class is now based on S3 classes. Because of this, it plays much nicer with the magrittr and related packages. There are now `str` and `summary` methods defined for `libbi` objects, and `sample`, `filter` and `optimise` commands for running LibBi